requirements.txt
langchain
openai>=1.0.0
chromadb
streamlit
googletrans==4.0.0-rc1
PyPDF2
tiktoken
python-dotenv
pdf2image
pytesseract
Pillow
python-docx
fitz
PyMuPDF
langdetect
nltk
tiktoken
fastapi
uvicorn
python-multipart



